fix: extract cli validate to a separate file
